change the wrapped feature duration to monthly summary
