+++ lib/screens/wrapped_intro_screen.dart
@@ -1,12 +1,58 @@
+import 'package:callstats/screens/wrapped_screen.dart';
 import 'package:flutter/material.dart';
 
-class WrappedIntroScreend extends StatelessWidget {
+class WrappedIntroScreen extends StatelessWidget {
   static const routeName = '/wrapped-intro';
 
-  const WrappedIntroScreend({super.key});
+  const WrappedIntroScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold();
+    return Scaffold(
+      body: Container(
+        decoration: const BoxDecoration(
+          gradient: LinearGradient(
+            colors: [Colors.blueAccent, Colors.purpleAccent],
+            begin: Alignment.topLeft,
+            end: Alignment.bottomRight,
+          ),
+        ),
+        child: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Center(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                const Text(
+                  'Welcome to your last 30 days Wrapped!',
+                  style: TextStyle(
+                    fontSize: 32,
+                    fontWeight: FontWeight.bold,
+                    color: Colors.white,
+                  ),
+                  textAlign: TextAlign.center,
+                ),
+                const SizedBox(height: 20),
+                const Text(
+                  'Discover new insights and enjoy your experience.',
+                  style: TextStyle(fontSize: 18, color: Colors.white),
+                  textAlign: TextAlign.center,
+                ),
+                const SizedBox(height: 40),
+                ElevatedButton(
+                  onPressed: () {
+                    Navigator.of(context).pop();
+                    Navigator.of(context).pushNamed(
+                      WrappedScreen.routeName,
+                    );
+                  },
+                  child: const Text('Get Started'),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
   }
 }
